fix(ci): gif dimensions guard, site-test-connection spy, extract buildPatch to pure ts
- image-dimensions.ts: lower global minimum bytes from 24 to 10 so
  GIF headers are not rejected before readGif runs
- site-test-connection.test.ts: replace originalGetMe (not a spy)
  with vi.mocked(wp.wpGetMe) in the input-validation assertion
- lib/platform/brand/build-patch.ts: extract FormState + buildPatch
  from CustomerBrandProfileEditor.tsx into a plain TS module so
  Vitest can import them without JSX parse errors (tsconfig jsx:preserve)
- CustomerBrandProfileEditor.tsx: import both from the new module;
  remove duplicate local definitions
- customer-brand-profile-editor-buildpatch.test.ts: update import path
